requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
dnspython==2.6.1
email_validator==2.1.1
idna==3.7
psycopg2==2.9.9
pydantic==2.7.2
pydantic_core==2.18.3
sqlparse==0.5.0
typing_extensions==4.12.0
